+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Pending additions
 ## New features
-- [#680](https://github.com/helmholtz-analytics/heat/pull/680) Introduction of property larray
+- [#680](https://github.com/helmholtz-analytics/heat/pull/680) New property: larray
+- [#683](https://github.com/helmholtz-analytics/heat/pull/683) New properties: nbytes, gnbytes, lnbytes
 ### Manipulations
 ### Statistical Functions
 ### Linear Algebra